把新加入蛇，重构到一个类。
