+++ src/main/java/ice/games/snake/OnopenProccesor.java
@@ -1,7 +1,6 @@
 package ice.games.snake;
 
 import java.io.IOException;
-import java.util.Iterator;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import org.atmosphere.cpr.AtmosphereResource;
@@ -23,22 +22,7 @@ public class OnopenProccesor {
 		Snake snake = new Snake(id, resource);
 		resource.session().setAttribute("snake", snake);
 
-		sendPlayingSnakeInfoToNew(snake);
-
 		snakeManager.addSnake(snake);
 	}
 
-	private void sendPlayingSnakeInfoToNew(Snake snake) {
-		Snake snaketemp;
-		StringBuilder sb = new StringBuilder();
-		for (Iterator<Snake> iterator = snakeManager.getPlayingSnakes().iterator(); iterator.hasNext();) {
-			snaketemp = iterator.next();
-			sb.append(String.format("{id: %d, color: '%s'}", Integer.valueOf(snaketemp.getId()),
-					snaketemp.getHexColor()));
-			if (iterator.hasNext()) {
-				sb.append(',');
-			}
-		}
-		snake.sendMessage(String.format("{'type': 'playinginfo','data':[%s]}", sb.toString()));
-	}
 }
